fix CI

+++ onchain/contracts/stello_pay_contract/src/enterprise.rs
@@ -576,7 +576,7 @@ impl HRWorkflowManager {
             workflow.employer,
             String::from_str(env, "InProgress"),
             completed_tasks,
-            workflow.checklist.len() as u32,
+            workflow.checklist.len(),
             current_time,
         );
